requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
flask-talisman==1.1.0
greenlet==3.0.0
gunicorn==21.2.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
Pygments==2.17.2
python-dotenv==1.0.0
rich==13.7.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
wrapt==1.16.0
